%% $OpenXM$
\name{hgm-package}
\alias{hgm-package}
\alias{hgm}
\docType{package}
\title{
HGM
}
\description{
The holonomic gradient method (hgm) gives a way to evaluate normalization
  constants of unnormalized probability distributions by utilizing holonomic 
  system of differential equations. The holonomic gradient descent gives a method
  to find maximal likelihood estimates by utilizing the hgm.
}
\details{
\tabular{ll}{
Package: \tab hgm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-07\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
More details.
}
\author{
Nobuki Takayama (takayama@math.kobe-u.ac.jp)
}
\references{
 \url{http://www.openxm.org}
}
\keyword{ package }
\seealso{
\code{\link{hgm.so3nc}}
}
\examples{
\dontrun{
example(hgm.so3nc)
}
}
